rename flags file and update imports
